chore: types and utils restructure
fix: linter config and errors
refactor: types modulariztation
chore: check workflow step
