+++ src/index.ts
@@ -1,5 +1,6 @@
 import { FileManagerBase } from './fileManager';
 
+export * from './types';
 export * from './utils';
 export * from './eventEmitter';
