+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export type { DriveInfo, FileInfo, ShareItem } from './info';
+export { FileStatus } from './info';
+export type { BrowserUploadOptions, FileInfoOptions, NodeUploadOptions, UploadProgress } from './utils';
+export type { FileManager } from './fileManager';
