+++ src/utils/common.ts
@@ -11,10 +11,11 @@ import {
   Topic,
 } from '@ethersphere/bee-js';
 
-import { FEED_INDEX_ZERO, SWARM_ZERO_ADDRESS } from './constants';
-import { FeedResultWithIndex, FeedPayloadResult, WrappedUploadResult } from './types';
-import { FileInfoError } from './errors';
+import { FeedPayloadResult, FeedResultWithIndex, WrappedUploadResult } from '../types/utils';
+
 import { assertWrappedUploadResult } from './asserts';
+import { FEED_INDEX_ZERO, SWARM_ZERO_ADDRESS } from './constants';
+import { FileInfoError } from './errors';
 
 export async function getFeedData(
   bee: Bee,
@@ -53,6 +54,7 @@ export async function getFeedData(
   }
 }
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function isNotFoundError(error: any): boolean {
   return error.stack?.includes('404') || error.message?.includes('Not Found') || error.message?.includes('404');
 }
@@ -101,6 +103,7 @@ export async function settlePromises<T>(promises: Promise<T>[], cb: (value: T) =
 export async function fetchStamp(bee: Bee, batchId: string | BatchId): Promise<PostageBatch | undefined> {
   try {
     return (await bee.getPostageBatches()).find((s) => s.batchID.toString() === batchId.toString());
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
   } catch (error: any) {
     console.error(`Failed to fetch stamp: ${error.message || error}`);
     return;
