+++ src/utils/index.ts
@@ -1,16 +1,5 @@
-export type {
-  DriveInfo,
-  FileInfo,
-  FileManager,
-  FileInfoOptions,
-  BrowserUploadOptions,
-  NodeUploadOptions,
-  ShareItem,
-  UploadProgress,
-} from './types';
-export { FileStatus } from './types';
 export { FileManagerEvents } from './events';
-export { ADMIN_STAMP_LABEL } from './constants';
+export { ADMIN_STAMP_LABEL, FILEMANAGER_STATE_TOPIC } from './constants';
 export {
   BeeVersionError,
   DriveError,
