+++ tests/mockHelpers.ts
@@ -26,9 +26,9 @@ import { BEE_URL, DEFAULT_MOCK_SIGNER } from './utils';
 
 import { EventEmitter } from '@/eventEmitter/eventEmitter';
 import { FileManagerBase } from '@/fileManager';
+import { DriveInfo, FileInfo } from '@/types';
+import { FileManagerEvents } from '@/utils';
 import { ADMIN_STAMP_LABEL, SWARM_ZERO_ADDRESS } from '@/utils/constants';
-import { FileManagerEvents } from '@/utils/events';
-import { DriveInfo, FileInfo } from '@/utils/types';
 
 export const MOCK_BATCH_ID = 'ee0fec26fdd55a1b8a777cc8c84277a1b16a7da318413fbd4cc4634dd93a2c51';
